requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch == 2.9.1+cu126
torchvision == 0.24.1+cu126
torchaudio == 2.9.1+cu126

PyQt6 == 6.10.2
kornia == 0.8.2
einops == 0.8.1
PyOpenGL == 3.1.10
requests == 2.32.5
scipy == 1.17.0
librosa == 0.11.0
SpoutGL == 0.1.1
ninja == 1.13.0